+++ src/Backend/Core/Engine/Form.php
@@ -8,6 +8,7 @@ use SpoonFormButton;
 use SpoonFormFile;
 use SpoonFormTextarea;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
+use function Symfony\Component\String\s;
 
 /**
  * This is our extended version of \SpoonForm
@@ -40,7 +41,7 @@ class Form extends \Common\Core\Form
         $url = BackendModel::getContainer()->get('url');
 
         parent::__construct(
-            $name ?? \SpoonFilter::toCamelCase($url->getModule() . '_' . $url->getAction(), '_', true),
+            $name ?? s($url->getModule() . ' ' . $url->getAction())->camel()->toString(),
             $action,
             $method ?? 'post',
             null,