+++ src/Backend/Modules/Pages/Actions/Edit.php
@@ -20,6 +20,7 @@ use Backend\Modules\Profiles\Engine\Model as BackendProfilesModel;
 use ForkCMS\Utility\Thumbnails;
 use SpoonFormHidden;
 use Symfony\Component\HttpFoundation\Response;
+use function Symfony\Component\String\s;
 
 /**
  * This is the edit-action, it will display a form to update an item
@@ -189,8 +190,8 @@ class Edit extends BackendBaseActionEdit
         // set headers
         $this->dgDrafts->setHeaderLabels(
             [
-                 'user_id' => \SpoonFilter::ucfirst(BL::lbl('By')),
-                 'edited_on' => \SpoonFilter::ucfirst(BL::lbl('LastEditedOn')),
+                 'user_id' => s(BL::lbl('By'))->title()->toString(),
+                 'edited_on' => s(BL::lbl('LastEditedOn'))->title()->toString(),
             ]
         );
 
@@ -365,7 +366,7 @@ class Edit extends BackendBaseActionEdit
 
             foreach ($permissions as $permission => $attributes) {
                 $values[] = [
-                    'label' => BL::msg(\SpoonFilter::toCamelCase('allow_' . $permission)),
+                    'label' => BL::msg(s('allow_' . $permission)->replace('_', ' ')->camel()->title()->toString()),
                     'value' => $permission,
                     'attributes' => $attributes,
                 ];
@@ -511,15 +512,15 @@ class Edit extends BackendBaseActionEdit
             $redirectValue = 'external';
         }
         $redirectValues = [
-            ['value' => 'none', 'label' => \SpoonFilter::ucfirst(BL::lbl('None'))],
+            ['value' => 'none', 'label' => s(BL::lbl('None'))->title()->toString()],
             [
                 'value' => 'internal',
-                'label' => \SpoonFilter::ucfirst(BL::lbl('InternalLink')),
+                'label' => s(BL::lbl('InternalLink'))->title()->toString(),
                 'variables' => ['isInternal' => true],
             ],
             [
                 'value' => 'external',
-                'label' => \SpoonFilter::ucfirst(BL::lbl('ExternalLink')),
+                'label' => s(BL::lbl('ExternalLink'))->title()->toString(),
                 'variables' => ['isExternal' => true],
             ],
         ];
@@ -595,8 +596,8 @@ class Edit extends BackendBaseActionEdit
         // set headers
         $this->dgRevisions->setHeaderLabels(
             [
-                 'user_id' => \SpoonFilter::ucfirst(BL::lbl('By')),
-                 'edited_on' => \SpoonFilter::ucfirst(BL::lbl('LastEditedOn')),
+                 'user_id' => s(BL::lbl('By'))->title()->toString(),
+                 'edited_on' => s(BL::lbl('LastEditedOn'))->title()->toString(),
             ]
         );