+++ src/Backend/Modules/Settings/Actions/Index.php
@@ -4,13 +4,13 @@ namespace Backend\Modules\Settings\Actions;
 
 use ForkCMS\Privacy\ConsentDialog;
 use ForkCMS\Utility\Akismet;
-use SpoonFilter;
 use Backend\Core\Engine\Base\ActionIndex as BackendBaseActionIndex;
 use Backend\Core\Engine\Form as BackendForm;
 use Backend\Core\Language\Language as BL;
 use Backend\Core\Engine\Model as BackendModel;
 use Backend\Modules\Extensions\Engine\Model as BackendExtensionsModel;
 use Backend\Modules\Settings\Engine\Model as BackendSettingsModel;
+use function Symfony\Component\String\s;
 
 /**
  * This is the index-action (default), it will display the setting-overview
@@ -137,13 +137,6 @@ class Index extends BackendBaseActionIndex
             'form-control danger code',
             true
         );
-        $this->form->addTextarea(
-            'site_domains',
-            implode("\n", (array) $this->get('fork.settings')->get('Core', 'site_domains', $defaultDomains)),
-            'form-control code',
-            'form-control danger code'
-        );
-
         // ckfinder
         $this->form->addText(
             'ckfinder_license_name',
@@ -190,11 +183,11 @@ class Index extends BackendBaseActionIndex
         $googleRecaptchaVersions = [
             [
                 'value' => 'v2invisible',
-                'label' => SpoonFilter::ucfirst(BL::lbl('GoogleRecaptchaV2Invisible')),
+                'label' => s(BL::lbl('GoogleRecaptchaV2Invisible'))->title()->toString(),
             ],
             [
                 'value' => 'v3',
-                'label' => SpoonFilter::ucfirst(BL::lbl('GoogleRecaptchaV3')),
+                'label' => s(BL::lbl('GoogleRecaptchaV3'))->title()->toString(),
             ],
         ];
         $this->form->addRadiobutton(
@@ -392,27 +385,6 @@ class Index extends BackendBaseActionIndex
                 }
             }
 
-            // domains filled in
-            if ($this->form->getField('site_domains')->isFilled()) {
-                // split on newlines
-                $domains = explode("\n", trim($this->form->getField('site_domains')->getValue()));
-
-                // loop domains
-                foreach ($domains as $domain) {
-                    // strip funky stuff
-                    $domain = trim(str_replace(['www.', 'http://', 'https://'], '', $domain));
-
-                    // invalid URL
-                    if (!SpoonFilter::isURL('http://' . $domain)) {
-                        // set error
-                        $this->form->getField('site_domains')->setError(BL::err('InvalidDomain'));
-
-                        // stop looping domains
-                        break;
-                    }
-                }
-            }
-
             if ($this->form->getField('ckfinder_image_max_width')->isFilled()) {
                 $this->form->getField(
                     'ckfinder_image_max_width'
@@ -577,23 +549,6 @@ class Index extends BackendBaseActionIndex
                 $this->get('fork.settings')->set('Core', 'redirect_languages', $redirectLanguages);
 
                 // domains may not contain www, http or https. Therefor we must loop and create the list of domains.
-                $siteDomains = [];
-
-                // domains filled in
-                if ($this->form->getField('site_domains')->isFilled()) {
-                    // split on newlines
-                    $domains = explode("\n", trim($this->form->getField('site_domains')->getValue()));
-
-                    // loop domains
-                    foreach ($domains as $domain) {
-                        // strip funky stuff
-                        $siteDomains[] = trim(str_replace(['www.', 'http://', 'https://'], '', $domain));
-                    }
-                }
-
-                // save domains
-                $this->get('fork.settings')->set('Core', 'site_domains', $siteDomains);
-
                 // privacy
                 $this->get('fork.settings')->set(
                     'Core',