+++ tests/data/test_db.sql
@@ -2353,7 +2353,6 @@ VALUES
 	('Core','akismet_key','s:0:\"\";'),
 	('Core','google_maps_key','s:0:\"\";'),
 	('Core','max_num_revisions','i:20;'),
-	('Core','site_domains','a:1:{i:0;s:13:\"fork.dev:8088\";}'),
 	('Core','site_html_head','s:0:\"\";'),
 	('Core','site_html_start_of_body','s:0:\"\";'),
 	('Core','site_html_end_of_body','s:0:\"\";'),